+++ notebooks/py_scripts/07-downstream-analysis-optional.py
@@ -7,9 +7,9 @@
 #       format_version: '1.5'
 #       jupytext_version: 1.14.1
 #   kernelspec:
-#     display_name: venv-nwb
+#     display_name: Python 3.9.13 ('ele')
 #     language: python
-#     name: venv-nwb
+#     name: python3
 # ---
 
 # + [markdown] tags=[]
@@ -24,14 +24,12 @@
 import os
 # change to the upper level folder to detect dj_local_conf.json
 if os.path.basename(os.getcwd())=='notebooks': os.chdir('..')
-assert os.path.basename(os.getcwd())=='workflow-calcium-imaging', (
-    "Please move to the workflow directory")
 # We'll be working with long tables, so we'll make visualization easier with a limit
 import datajoint as dj; dj.config['display.limit']=10
 
 # Next, we populate the python namespace with the required schemas
 
-from workflow_calcium_imaging.pipeline import session, imaging, trial, event
+from workflow_calcium_imaging.pipeline import session, imaging, trial, event, analysis
 
 # + [markdown] jp-MarkdownHeadingCollapsed=true tags=[]
 # ## Trial and Event schemas
@@ -78,9 +76,6 @@ event.AlignmentEvent()
 
 # + [markdown] tags=[]
 # # Event-aligned trialized calcium activity
-# -
-
-from workflow_calcium_imaging import analysis
 
 # + [markdown] jp-MarkdownHeadingCollapsed=true tags=[]
 # ### Analysis
@@ -97,7 +92,7 @@ imaging.Activity()
 # We'll isolate the scan of interest with the following key:
 
 ca_activity_key = (imaging.Activity & {'subject': 'subject3', 'scan_id': 0}
-                  ).fetch1('KEY')
+                  ).fetch('KEY')[0]
 
 # Here, we can see all trials for this scan:
 
@@ -188,7 +183,7 @@ imaging.Segmentation.Mask & 'mask<3'
 from workflow_calcium_imaging import analysis
 from workflow_calcium_imaging.pipeline import session, imaging, trial, event
 ca_activity_key = (imaging.Activity & {'subject': 'subject3', 'scan_id': 0}
-                  ).fetch1('KEY')
+                  ).fetch('KEY')[0]
 alignment_key = (event.AlignmentEvent & 'alignment_name = "center_button"'
                 ).fetch1('KEY')
 alignment_condition_ctrl = {**ca_activity_key, **alignment_key,