+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 Observes [Semantic Versioning](https://semver.org/spec/v2.0.0.html) standard and 
 [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) convention.
 
+## 0.1.1 - 2022-09-18
++ Add - Code of conduct
++ Add - CICD workflows
++ Add - sbxreader requirement
++ Update - notebooks for codebook deploymnet
++ Update - csv session datetime for codebook data
+
+
 ## 0.1.0 - 2022-06-29
 
 + Add - Suite2p triggering pytest